one to many db, correct user edit, delete post

+++ posts/bluerprints.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, render_template, flash, request, redirect
 from flask.helpers import url_for
-from flask_login import login_required
+from flask_login import login_required, current_user
 
 from db import db
 from .models import Posts
@@ -22,15 +22,16 @@ def posts(id):
 
 
 @post.route('/add-post', methods=['GET', 'POST'])
+@login_required
 def add_post():
     form = PostForm()
 
     if form.validate_on_submit():
-        post = Posts(title=form.title.data, content=form.content.data,
-                     autor=form.autor.data, slug=form.slug.data)
+        poster = current_user.id
+        post = Posts(title=form.title.data, poster_id=poster,
+                     content=form.content.data, slug=form.slug.data)
         form.title.data = ''
         form.content.data = ''
-        form.autor.data = ''
         form.slug.data = ''
 
         db.session.add(post)
@@ -45,19 +46,19 @@ def add_post():
 def edit_post(id):
     form = PostForm()
     post = Posts.query.get_or_404(id)
-    if form.validate_on_submit():
-        post.title = form.title.data
-        post.content = form.content.data
-        post.autor = form.autor.data
-        post.slug = form.slug.data
+    id = current_user.id
+    if id == post.poster.id:
+        if form.validate_on_submit():
+            post.title = form.title.data
+            post.content = form.content.data
+            post.slug = form.slug.data
 
-        db.session.add(post)
-        db.session.commit()
-        flash('Данные обновленны.')
-        return redirect(url_for('post.posts', id=post.id))
+            db.session.add(post)
+            db.session.commit()
+            flash('Данные обновленны.')
+            return redirect(url_for('post.posts', id=post.id))
     form.title.data = post.title
     form.content.data = post.content
-    form.autor.data = post.autor
     form.slug.data = post.slug
     return render_template('posts/edit.html', form=form, post=post)
 
@@ -66,12 +67,15 @@ def edit_post(id):
 @login_required
 def delete_post(id):
     to_delete = Posts.query.get_or_404(id)
-
-    try:
-        db.session.delete(to_delete)
-        db.session.commit()
-        flash('Статья удаленна')
-    except:
-        flash('Ошибка при удалении! Попробуй еще раз...')
+    id = current_user.id
+    if id == to_delete.poster.id:
+        try:
+            db.session.delete(to_delete)
+            db.session.commit()
+            flash('Статья удаленна')
+        except:
+            flash('Ошибка при удалении! Попробуй еще раз...')
+    else:
+        flash('Вы не можете удалить эту страницу!')
 
     return redirect(url_for('post.index'))
